Woof-3: Entire Overhaul of File structure  (#42)
* Woof-3 : documentation additions

* Woof-3 : final updates to documentation

* Woof-3 : Restructured the design

* Woof-3 : Refactored entire folder layout structure
