+++ pages/all/index.tsx
@@ -3,11 +3,17 @@ import { fetchArticles } from '@/fetchArticles';
 import {
   ArticleSectionDisplay,
   ArticlesProps,
-} from '../components/ArticleSectionDisplay';
+} from '../../components/body/ArticleSectionDisplay';
+import Woof_layout from '../layout';
+import { useRouter } from 'next/router';
 
 const All = ({ articles, allArticles }: ArticlesProps) => {
+  const route = useRouter();
+
   return (
-    <ArticleSectionDisplay articles={articles} allArticles={allArticles} />
+    <Woof_layout pageProps={articles} router={route.route}>
+      <ArticleSectionDisplay articles={articles} allArticles={allArticles} />
+    </Woof_layout>
   );
 };
