+++ pages/misc/index.tsx
@@ -3,17 +3,23 @@ import { fetchArticles } from '@/fetchArticles';
 import {
   ArticleSectionDisplay,
   ArticlesProps,
-} from '../components/ArticleSectionDisplay';
+} from '../../components/body/ArticleSectionDisplay';
 import { MISC } from '@/constants';
+import Woof_layout from '../layout';
+import { useRouter } from 'next/router';
 
 const Misc = ({ articles, allArticles }: ArticlesProps) => {
+  const route = useRouter();
+
   return (
+    <Woof_layout pageProps={articles} router={route.route}>
     <ArticleSectionDisplay
       articles={articles}
       allArticles={allArticles}
       section={MISC}
       color="misc-color"
     />
+    </Woof_layout>
   );
 };
