+++ pages/opinion/index.tsx
@@ -3,17 +3,24 @@ import { fetchArticles } from '@/fetchArticles';
 import {
   ArticleSectionDisplay,
   ArticlesProps,
-} from '../components/ArticleSectionDisplay';
+} from '../../components/body/ArticleSectionDisplay';
 import { OPINION } from '@/constants';
+import Woof_layout from '../layout';
+import { useRouter } from 'next/router';
 
 const Opinion = ({ articles, allArticles }: ArticlesProps) => {
+  const route = useRouter();
+
+
   return (
-    <ArticleSectionDisplay
-      articles={articles}
-      allArticles={allArticles}
-      section={OPINION}
-      color="opinion-color"
-    />
+    <Woof_layout pageProps={articles} router={route.route}>
+      <ArticleSectionDisplay
+        articles={articles}
+        allArticles={allArticles}
+        section={OPINION}
+        color="opinion-color"
+      />
+    </Woof_layout>
   );
 };
