+++ pages/sports/index.tsx
@@ -3,17 +3,23 @@ import { fetchArticles } from '@/fetchArticles';
 import {
   ArticleSectionDisplay,
   ArticlesProps,
-} from '../components/ArticleSectionDisplay';
-import { TROUBLESHOOTING } from '@/constants';
+} from '../../components/body/ArticleSectionDisplay';
+import { SPORTS } from '@/constants';
+import Woof_layout from '../layout';
+import { useRouter } from 'next/router';
 
-const Troubleshooting = ({ articles, allArticles }: ArticlesProps) => {
+const Sports = ({ articles, allArticles }: ArticlesProps) => {
+  const route = useRouter();
+
   return (
+    <Woof_layout pageProps={articles} router={route.route}>
     <ArticleSectionDisplay
       articles={articles}
       allArticles={allArticles}
-      section={TROUBLESHOOTING}
-      color="troubleshooting-color"
+      section={SPORTS}
+      color="sports-color"
     />
+    </Woof_layout>
   );
 };
 
@@ -22,9 +28,9 @@ export const getServerSideProps: GetServerSideProps<
 > = async () => {
   const allArticles = await fetchArticles();
   const articles = allArticles.filter((article) =>
-    article.sections.includes(TROUBLESHOOTING),
+    article.sections.includes(SPORTS),
   );
   return { props: { articles, allArticles } };
 };
 
-export default Troubleshooting;
+export default Sports;
